issue #3: Rename modules
