+++ reactivexcomponent/reactivexcomponent/__init__.py
@@ -1,9 +1,9 @@
 import logging
 
-from reactivexcomponent.xcomponentAPI import XcomponentAPI
+from reactivexcomponent.xcomponent_api import XcomponentAPI
 from reactivexcomponent.communication import publisher
-from reactivexcomponent.communication import xcConnection
-from reactivexcomponent.communication import xcSession
+from reactivexcomponent.communication import xc_connection
+from reactivexcomponent.communication import xc_session
 from reactivexcomponent.configuration import serializer
 
 __version__ = '1.0.0'
