+++ reactivexcomponent/reactivexcomponent/communication/xc_connection.py
@@ -1,7 +1,10 @@
-from reactivexcomponent.communication.xcSession import XcSession
+from reactivexcomponent.communication.xc_session import XcSession
 
 
 class XcConnection:
+
+    def __init__(self):
+        self.session = None
 
     def create_connection(self, xc_api, server_url, callback):
         self.session = XcSession()
